refactor: split vectorise_graph in two scripts
feat: fix bugs in vectorise gpao version

fix: allow blankspace for paths in command lines

fix: allow blankspace for paths in command lines

fix: bad blankspace in command lines

refactor: clean code

fix: clean verbose

fix: verbose

fix: modifications following code review

fix: create correct graph when there are patches

fix: vectorise_graph help

fix: bug with default output path

fix: output is required for prep_vectorise
