--- scripts/modifyvrt.py
@@ -1,51 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Script permettant de passer d'un VRT 3 canaux 8 bits à un VRT monocanal 32 bits
-"""
-
-import argparse
-
-def read_args():
-    """Gestion des arguments"""
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("-i", "--input", required=True, help="input cache folder")
-    parser.add_argument("-o", "--output", help="output folder")
-
-    if args.verbose >= 1:
-        print("\nArguments: ", args)
-
-    return args
-
-
-args = read_args()
-
-
-with open(args.input, 'r') as f:
-    lines = f.readlines()
-with open(args.output, 'w') as f:
-    for line in lines:
-        # on ecrit le code python au bon endroit dans le VRT
-        if 'band="1"' in line:
-            f.write('\t<VRTRasterBand dataType="Int32" band="1" subClass="VRTDerivedRasterBand">\n')
-            f.write('\t<PixelFunctionType>color_to_int32</PixelFunctionType>\n')
-            f.write('\t<PixelFunctionLanguage>Python</PixelFunctionLanguage>\n')
-            f.write('\t<PixelFunctionCode>\n')
-            f.write('<![CDATA[\n')
-            f.write('import numpy as np\n')
-            f.write('def color_to_int32(in_ar, out_ar, xoff, yoff, xsize, ysize, \
-                    raster_xsize, raster_ysize, buf_radius, gt, **kwargs):\n')
-            f.write('\tout_ar[:] = in_ar[0] + 256 * in_ar[1] + 256 * 256 * in_ar[2]\n')
-            f.write(']]>\n')
-            f.write('\t</PixelFunctionCode>\n')
-        elif 'band="2"' in line:
-            pass
-        elif 'band="3"' in line:
-            pass
-        elif '</VRTRaster' in line:
-            pass
-        elif '<OverviewList' in line:
-            f.write('\t</VRTRasterBand>\n')
-            f.write(line)
-        else:
-            f.write(line)
